+++ cw-orch-daemon/src/core.rs
@@ -79,6 +79,12 @@ impl DaemonAsync {
     pub fn channel(&self) -> Channel {
         self.sender.grpc_channel.clone()
     }
+
+    /// Flushes all the state related to the current chain
+    /// Only works on Local networks
+    pub fn flush_state(&self) -> Result<(), DaemonError> {
+        self.state.flush()
+    }
 }
 
 impl ChainState for DaemonAsync {