+++ cw-orch-daemon/src/error.rs
@@ -122,6 +122,8 @@ pub enum DaemonError {
     QuerierNeedRuntime,
     #[error(transparent)]
     Instantiate2Error(#[from] Instantiate2AddressError),
+    #[error("Error opening file {0},err: ({1})")]
+    OpenFile(String, String),
     #[error("State file {0} already locked, use another state file or clone daemon which holds the lock")]
     StateAlreadyLocked(String),
 }