+++ cw-orch-daemon/src/json_lock.rs
@@ -97,8 +97,8 @@ impl Drop for JsonLockedState {
 }
 
 pub fn read(filename: &String) -> Result<Value, DaemonError> {
-    let file =
-        File::open(filename).unwrap_or_else(|_| panic!("File should be present at {}", filename));
+    let file = File::open(filename)
+        .map_err(|err| DaemonError::OpenFile(filename.to_string(), err.to_string()));
     let json: serde_json::Value = from_reader(file)?;
     Ok(json)
 }