+++ cw-orch/src/prelude.rs
@@ -36,10 +36,6 @@ pub use crate::environment::{ChainInfo, ChainInfoOwned};
 // Mock for testing
 pub use crate::mock::{Mock, MockBech32};
 
-// OsmosisTestTube for testing
-#[cfg(feature = "osmosis-test-tube")]
-pub use crate::osmosis_test_tube::OsmosisTestTube;
-
 // error
 pub use crate::error::CwOrchError;
 
@@ -72,3 +68,16 @@ pub use cw_orch_traits::*;
 
 #[cfg(feature = "snapshot-testing")]
 pub use crate::take_storage_snapshot;
+
+#[cfg(feature = "interchain")]
+/// Everything that concerns interchain capabilities
+pub mod interchain {
+    pub use cw_orch_interchain_core::{IbcQueryHandler, InterchainEnv};
+    pub use cw_orch_interchain_daemon::{
+        ChannelCreationValidator, ChannelCreator, DaemonInterchainEnv,
+    };
+    pub use cw_orch_interchain_mock::{MockBech32InterchainEnv, MockInterchainEnv};
+    pub use cw_orch_starship::Starship;
+}
+#[cfg(feature = "interchain")]
+pub use interchain::*;