+++ packages/macros/cw-orch-fns-derive/src/execute_fns.rs
@@ -1,0 +1,8 @@
+pub fn payable(v: &syn::Variant) -> bool {
+    for attr in &v.attrs {
+        if attr.path.segments.len() == 1 && attr.path.segments[0].ident == "payable" {
+            return true;
+        }
+    }
+    false
+}