+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+- Added a test to make sure the derive macros stay compatible with new cw-orch versions
+- Changed the derive macros import from cw_orch to cw_orch_core. This allows changing the cw-orch API without breaking the derive macros.
+- Cw-orch mock env info doesn't error when using chain ids that don't match the `osmosis-1` pattern
+- Add interchain capabilites as well as clone-testing
+- Bumped MSRV to 1.73 because of dependency `cosmwasm-vm@1.5.5`
+- Remove `impl_into`, the old `impl_into` behavior is now the default behavior
+- EXCITING FEATURE : Added an item and a map query method to be able to query cw-storage-plus structure outside of contracts easily
+- Add `flush_state` method for Local Chain Daemons
+- cw-orch-interchain now errors on checking transactions for IBC packets if NO packets were found
+
+### Breaking
+
+- Daemon : Changed return types on daemon queriers to match Cosmwasm std types
+- Cw-orch : Separate osmosis test tube from cw-orch. Its not available in its own crate `cw-orch-osmosis-test-tube`
+
+## 0.22.0
+
 - Updated osmosis test tube to 24.0.1 ,that avoids re-compiling osmosis test tube
 - Added `balance` query at the root of QueryHandler
 - Added DaemonBuilder configuration for grpc url and fee overwriting
@@ -14,6 +31,7 @@
 ## 0.21.2
 
 - Allow cw-orch wasm compilation without features
+- Transaction Response now inspects logs and events to find matching events.
 
 ## 0.21.1